+++ Kokoro/Features/Actions/ConditionalActionsImpl.cs
@@ -181,13 +181,13 @@ internal sealed class ConditionalActionHasStatusExpression : IKokoroApi.IConditi
 	}
 
 	public string GetTooltipDescription(State state, Combat? combat)
-		=> string.Format(I18n.ConditionalHasStatusDescription, $"<c=status>{Status.GetLocName().ToUpper()}</c>");
+		=> string.Format(TargetPlayer ? I18n.ConditionalHasStatusDescription : I18n.ConditionalEnemyHasStatusDescription, $"<c=status>{Status.GetLocName().ToUpper()}</c>");
 
 	public List<Tooltip> GetTooltips(State state, Combat? combat)
 	{
 		var ship = TargetPlayer ? state.ship : combat?.otherShip;
 		var amount = ship?.Get(Status) ?? 1;
-		return new() { new TTGlossary($"status.{Status.Key()}", amount) };
+		return [new TTGlossary($"status.{Status.Key()}", amount)];
 	}
 }
 
@@ -336,6 +336,16 @@ internal sealed class ConditionalActionEquation : IKokoroApi.IConditionalActionA
 				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationStateLessThanOrEqualDescription,
 				_ => throw new ArgumentException()
 			},
+			IKokoroApi.IConditionalActionApi.EquationStyle.EnemyState => Operator switch
+			{
+				IKokoroApi.IConditionalActionApi.EquationOperator.Equal => I18n.ConditionalEquationEnemyStateEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.NotEqual => I18n.ConditionalEquationEnemyStateNotEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThan => I18n.ConditionalEquationEnemyStateGreaterThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThan => I18n.ConditionalEquationEnemyStateLessThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThanOrEqual => I18n.ConditionalEquationEnemyStateGreaterThanOrEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationEnemyStateLessThanOrEqualDescription,
+				_ => throw new ArgumentException()
+			},
 			IKokoroApi.IConditionalActionApi.EquationStyle.Possession => Operator switch
 			{
 				IKokoroApi.IConditionalActionApi.EquationOperator.Equal => I18n.ConditionalEquationPossessionEqualDescription,
@@ -346,6 +356,16 @@ internal sealed class ConditionalActionEquation : IKokoroApi.IConditionalActionA
 				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationPossessionLessThanOrEqualDescription,
 				_ => throw new ArgumentException()
 			},
+			IKokoroApi.IConditionalActionApi.EquationStyle.EnemyPossession => Operator switch
+			{
+				IKokoroApi.IConditionalActionApi.EquationOperator.Equal => I18n.ConditionalEquationEnemyPossessionEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.NotEqual => I18n.ConditionalEquationEnemyPossessionNotEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThan => I18n.ConditionalEquationEnemyPossessionGreaterThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThan => I18n.ConditionalEquationEnemyPossessionLessThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThanOrEqual => I18n.ConditionalEquationEnemyPossessionGreaterThanOrEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationEnemyPossessionLessThanOrEqualDescription,
+				_ => throw new ArgumentException()
+			},
 			IKokoroApi.IConditionalActionApi.EquationStyle.PossessionComparison => Operator switch
 			{
 				IKokoroApi.IConditionalActionApi.EquationOperator.Equal => I18n.ConditionalEquationPossessionComparisonEqualDescription,
@@ -354,6 +374,26 @@ internal sealed class ConditionalActionEquation : IKokoroApi.IConditionalActionA
 				IKokoroApi.IConditionalActionApi.EquationOperator.LessThan => I18n.ConditionalEquationPossessionComparisonLessThanDescription,
 				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThanOrEqual => I18n.ConditionalEquationPossessionComparisonGreaterThanOrEqualDescription,
 				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationPossessionComparisonLessThanOrEqualDescription,
+				_ => throw new ArgumentException()
+			},
+			IKokoroApi.IConditionalActionApi.EquationStyle.EnemyPossessionComparison => Operator switch
+			{
+				IKokoroApi.IConditionalActionApi.EquationOperator.Equal => I18n.ConditionalEquationEnemyPossessionComparisonEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.NotEqual => I18n.ConditionalEquationEnemyPossessionComparisonNotEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThan => I18n.ConditionalEquationEnemyPossessionComparisonGreaterThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThan => I18n.ConditionalEquationEnemyPossessionComparisonLessThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThanOrEqual => I18n.ConditionalEquationEnemyPossessionComparisonGreaterThanOrEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationEnemyPossessionComparisonLessThanOrEqualDescription,
+				_ => throw new ArgumentException()
+			},
+			IKokoroApi.IConditionalActionApi.EquationStyle.ToEnemyPossessionComparison => Operator switch
+			{
+				IKokoroApi.IConditionalActionApi.EquationOperator.Equal => I18n.ConditionalEquationToEnemyPossessionComparisonEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.NotEqual => I18n.ConditionalEquationToEnemyPossessionComparisonNotEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThan => I18n.ConditionalEquationToEnemyPossessionComparisonGreaterThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThan => I18n.ConditionalEquationToEnemyPossessionComparisonLessThanDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.GreaterThanOrEqual => I18n.ConditionalEquationToEnemyPossessionComparisonGreaterThanOrEqualDescription,
+				IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual => I18n.ConditionalEquationToEnemyPossessionComparisonLessThanOrEqualDescription,
 				_ => throw new ArgumentException()
 			},
 			_ => throw new ArgumentException()